Request backups when final event comes through.
